clean up and rename components
