+++ client/app/adminSignInAndSignUp/page.jsx
@@ -8,8 +8,8 @@ import {
 } from "react-firebase-hooks/auth";
 import { auth } from "../firebase/config";
 import { useRouter } from "next/navigation";
-import AdminSignIn from "../_components/AdminSignIn/AdminSignIn";
-import AdminSignUp from "../_components/AdminSignIn/AdminSignUp";
+import AdminSignIn from "../_components/SignInForms/SignIn";
+import AdminSignUp from "../_components/SignInForms/SignUp";
 
 export default function AdminSignInAndSignUp() {
   const [user, loading, error] = useAuthState(auth);
@@ -20,57 +20,6 @@ export default function AdminSignInAndSignUp() {
       router.push("/admin/dashboard");
     }
   }, [user]);
-
-  // useEffect(() => {
-  //   if (signUpError) {
-  //     if (signUpError.message.includes("email-already-in-use")) {
-  //       setSignUpAlert("Email already in use");
-  //     } else {
-  //       setSignUpAlert("");
-  //     }
-  //   }
-  // }, [signUpError]);
-
-  // useEffect(() => {
-  //   if (signInError) {
-  //     if (signInError.message.includes("invalid-credential")) {
-  //       setSignInAlert("Either email or password is incorrect");
-  //     } else {
-  //       setSignInAlert("");
-  //     }
-  //   }
-  // }, [signInError]);
-
-  // const signInFormSubmitHandler = async (e) => {
-  //   e.preventDefault();
-  //   try {
-  //     const res = await signInWithEmailAndPassword(signinEmail, signinPassword);
-  //     setSigninEmail("");
-  //     setSigninPassword("");
-  //     if (res) {
-  //       router.push("/admin/dashboard");
-  //     }
-  //   } catch (error) {
-  //     console.log("error", error);
-  //   }
-  // };
-
-  // const signUpFormSubmitHandler = async (e) => {
-  //   e.preventDefault();
-  //   try {
-  //     const res = await createUserWithEmailAndPassword(
-  //       signupEmail,
-  //       signupPassword
-  //     );
-  //     setSignupEmail("");
-  //     setSignupPassword("");
-  //     if (res) {
-  //       router.push("/admin/dashboard");
-  //     }
-  //   } catch (error) {
-  //     console.log("error", error);
-  //   }
-  // };
 
   if (loading || user) {
     return (
